+++ spark-yun-backend/src/main/java/com/isxcode/star/backend/module/engine/node/entity/EngineNodeEntity.java
@@ -30,13 +30,13 @@ public class EngineNodeEntity {
 
   private LocalDateTime checkDateTime;
 
-  private Integer allMemory;
+  private Double allMemory;
 
-  private Integer usedMemory;
+  private Double usedMemory;
 
-  private Integer allStorage;
+  private Double allStorage;
 
-  private Integer usedStorage;
+  private Double usedStorage;
 
   private String cpuPercent;
